* Removed warnings observed by GCC
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8558 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/diagraphica/attrconti.cpp
@@ -200,7 +200,7 @@ void AttrConti::moveValue(
 
 
 // --------------------------------------------------------
-void AttrConti::classifyEqualIntervals( const int &number )
+void AttrConti::classifyEqualIntervals( const size_t &number )
 // --------------------------------------------------------
 {
     if ( number == 0 )
@@ -245,7 +245,7 @@ void AttrConti::classifyEqualIntervals( const int &number )
 
 
 // ---------------------------------------------------
-void AttrConti::classifyQuantiles( const int &number )
+void AttrConti::classifyQuantiles( const size_t &number )
 // ---------------------------------------------------
 {
     if ( number == 0 )
@@ -291,7 +291,7 @@ void AttrConti::classifyQuantiles( const int &number )
 
 
 // ---------------------------------------------------------------
-void AttrConti::classifyMeanStandardDeviation( const int &number )
+void AttrConti::classifyMeanStandardDeviation( const size_t &number )
 // ---------------------------------------------------------------
 {
     if ( number == 0 )
